Module Sentry__.Capped_string_4k

type t = string

The Sentry API discards anything after the first 512 characters of most string attributes, so avoid sending them in the first place, so we can also avoid sending events that are too big. https://docs.sentry.io/accounts/quotas/#attributes-limits

val sexp_of_t : t -> Ppx_sexp_conv_lib.Sexp.t

val wrap : string -> t

val unwrap : t -> string