Module Sentry__Context

type t = {

mutable environment : string option;

mutable release : string option;

mutable server_name : string option;

tags : string Core_kernel.String.Table.t;

extra : Sentry__.Json.t Core_kernel.String.Table.t;

modules : string Core_kernel.String.Table.t;

breadcrumbs : Sentry.Breadcrumb.t Core_kernel.Queue.t;

}

val sexp_of_t : t -> Ppx_sexp_conv_lib.Sexp.t

val default : ?⁠max_breadcrumbs:int -> unit -> t

Returns a new context with default data from the environment and system calls.

val empty : ?⁠max_breadcrumbs:int -> unit -> t

Returns a new context with no tags or breadcrumbs. You probably want to use default () or copying the parent context in most cases.

val copy : t -> t

Copy a context so you can add / clear data without affecting the original

val merge_tags : (string * string) list -> t -> unit

merge_tags tags t merges the given tags into t

val merge_extra : (string * Sentry__.Json.t) list -> t -> unit

merge_extra extra t merges the given extra data into t

val merge_modules : (string * string) list -> t -> unit

merge_modules modules t merges the given module info into t

val add_breadcrumb : Sentry.Breadcrumb.t -> t -> unit

Add a breadcrumb to the context and remove older breadcrumbs