Module Sentry__Payloads_j

type uuid = Sentry__.Payloads_t.uuid

type platform = Sentry__.Payloads_t.platform

type capped_string_512 = Sentry__.Payloads_t.capped_string_512

type stackframe = Sentry__.Payloads_t.stackframe = {

filename : capped_string_512 option;

function_ : capped_string_512 option;

module_ : capped_string_512 option;

lineno : int option;

colno : int option;

abs_path : capped_string_512 option;

context_line : capped_string_512 option;

pre_context : capped_string_512 list option;

post_context : capped_string_512 list option;

in_app : bool option;

vars : (capped_string_512 * capped_string_512) list option;

package : capped_string_512 option;

platform : platform option;

}

type stacktrace = Sentry__.Payloads_t.stacktrace = {

frames : stackframe list;

}

type severity_level = Sentry__.Payloads_t.severity_level

type sdk_info = Sentry__.Payloads_t.sdk_info = {

name : capped_string_512;

version : capped_string_512;

integrations : capped_string_512 list option;

}

type response = Sentry__.Payloads_t.response = {

id : uuid;

}

type capped_string_4k = Sentry__.Payloads_t.capped_string_4k

type message = Sentry__.Payloads_t.message = {

message : capped_string_4k;

params : capped_string_512 list option;

formatted : capped_string_512 option;

}

type mechanism = Sentry__.Payloads_t.mechanism = {

type_ : capped_string_512;

description : capped_string_512 option;

help_link : capped_string_512 option;

handled : bool option;

data : (capped_string_512 * capped_string_512) list option;

}

type json = Yojson.Basic.json

type exception_value = Sentry__.Payloads_t.exception_value = {

type_ : capped_string_512;

value : capped_string_4k option;

module_ : capped_string_512 option;

thread_id : capped_string_512 option;

mechanism : mechanism option;

stacktrace : stacktrace option;

}

type exception_ = Sentry__.Payloads_t.exception_ = {

values : exception_value list;

}

type datetime = Sentry__.Payloads_t.datetime

type breadcrumb = Sentry__.Payloads_t.breadcrumb = {

timestamp : datetime;

type_ : capped_string_512 option;

message : capped_string_512 option;

data : (capped_string_512 * json) list option;

category : capped_string_512 option;

level : capped_string_512 option;

}

type event = Sentry__.Payloads_t.event = {

event_id : uuid;

timestamp : datetime;

logger : capped_string_512 option;

platform : platform;

sdk : sdk_info;

level : severity_level option;

culprit : capped_string_512 option;

server_name : capped_string_512 option;

release : capped_string_512 option;

tags : (capped_string_512 * capped_string_512) list option;

environment : capped_string_512 option;

modules : (capped_string_512 * capped_string_512) list option;

extra : (capped_string_512 * json) list option;

fingerprint : capped_string_512 list option;

exception_ : exception_ option;

message : message option;

breadcrumbs : breadcrumb list option;

}

val write_uuid : Bi_outbuf.t -> uuid -> unit

Output a JSON value of type uuid.

val string_of_uuid : ?⁠len:int -> uuid -> string

Serialize a value of type uuid into a JSON string.

parameter len

specifies the initial length of the buffer used internally. Default: 1024.

val read_uuid : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> uuid

Input JSON data of type uuid.

val uuid_of_string : string -> uuid

Deserialize JSON data of type uuid.

val write_platform : Bi_outbuf.t -> platform -> unit

Output a JSON value of type platform.

val string_of_platform : ?⁠len:int -> platform -> string

Serialize a value of type platform into a JSON string.

parameter len

specifies the initial length of the buffer used internally. Default: 1024.

val read_platform : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> platform

Input JSON data of type platform.

val platform_of_string : string -> platform

Deserialize JSON data of type platform.

val write_capped_string_512 : Bi_outbuf.t -> capped_string_512 -> unit

Output a JSON value of type capped_string_512.

val string_of_capped_string_512 : ?⁠len:int -> capped_string_512 -> string

Serialize a value of type capped_string_512 into a JSON string.

parameter len

specifies the initial length of the buffer used internally. Default: 1024.

val read_capped_string_512 : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> capped_string_512

Input JSON data of type capped_string_512.

val capped_string_512_of_string : string -> capped_string_512

Deserialize JSON data of type capped_string_512.

val write_stackframe : Bi_outbuf.t -> stackframe -> unit

Output a JSON value of type stackframe.

val string_of_stackframe : ?⁠len:int -> stackframe -> string

Serialize a value of type stackframe into a JSON string.

parameter len

specifies the initial length of the buffer used internally. Default: 1024.

val read_stackframe : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> stackframe

Input JSON data of type stackframe.

val stackframe_of_string : string -> stackframe

Deserialize JSON data of type stackframe.

val write_stacktrace : Bi_outbuf.t -> stacktrace -> unit

Output a JSON value of type stacktrace.

val string_of_stacktrace : ?⁠len:int -> stacktrace -> string

Serialize a value of type stacktrace into a JSON string.

parameter len

specifies the initial length of the buffer used internally. Default: 1024.

val read_stacktrace : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> stacktrace

Input JSON data of type stacktrace.

val stacktrace_of_string : string -> stacktrace

Deserialize JSON data of type stacktrace.

val write_severity_level : Bi_outbuf.t -> severity_level -> unit

Output a JSON value of type severity_level.

val string_of_severity_level : ?⁠len:int -> severity_level -> string

Serialize a value of type severity_level into a JSON string.

parameter len

specifies the initial length of the buffer used internally. Default: 1024.

val read_severity_level : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> severity_level

Input JSON data of type severity_level.

val severity_level_of_string : string -> severity_level

Deserialize JSON data of type severity_level.

val write_sdk_info : Bi_outbuf.t -> sdk_info -> unit

Output a JSON value of type sdk_info.

val string_of_sdk_info : ?⁠len:int -> sdk_info -> string

Serialize a value of type sdk_info into a JSON string.

parameter len

specifies the initial length of the buffer used internally. Default: 1024.

val read_sdk_info : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> sdk_info

Input JSON data of type sdk_info.

val sdk_info_of_string : string -> sdk_info

Deserialize JSON data of type sdk_info.

val write_response : Bi_outbuf.t -> response -> unit

Output a JSON value of type response.

val string_of_response : ?⁠len:int -> response -> string

Serialize a value of type response into a JSON string.

parameter len

specifies the initial length of the buffer used internally. Default: 1024.

val read_response : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> response

Input JSON data of type response.

val response_of_string : string -> response

Deserialize JSON data of type response.

val write_capped_string_4k : Bi_outbuf.t -> capped_string_4k -> unit

Output a JSON value of type capped_string_4k.

val string_of_capped_string_4k : ?⁠len:int -> capped_string_4k -> string

Serialize a value of type capped_string_4k into a JSON string.

parameter len

specifies the initial length of the buffer used internally. Default: 1024.

val read_capped_string_4k : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> capped_string_4k

Input JSON data of type capped_string_4k.

val capped_string_4k_of_string : string -> capped_string_4k

Deserialize JSON data of type capped_string_4k.

val write_message : Bi_outbuf.t -> message -> unit

Output a JSON value of type message.

val string_of_message : ?⁠len:int -> message -> string

Serialize a value of type message into a JSON string.

parameter len

specifies the initial length of the buffer used internally. Default: 1024.

val read_message : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> message

Input JSON data of type message.

val message_of_string : string -> message

Deserialize JSON data of type message.

val write_mechanism : Bi_outbuf.t -> mechanism -> unit

Output a JSON value of type mechanism.

val string_of_mechanism : ?⁠len:int -> mechanism -> string

Serialize a value of type mechanism into a JSON string.

parameter len

specifies the initial length of the buffer used internally. Default: 1024.

val read_mechanism : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> mechanism

Input JSON data of type mechanism.

val mechanism_of_string : string -> mechanism

Deserialize JSON data of type mechanism.

val write_json : Bi_outbuf.t -> json -> unit

Output a JSON value of type json.

val string_of_json : ?⁠len:int -> json -> string

Serialize a value of type json into a JSON string.

parameter len

specifies the initial length of the buffer used internally. Default: 1024.

val read_json : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> json

Input JSON data of type json.

val json_of_string : string -> json

Deserialize JSON data of type json.

val write_exception_value : Bi_outbuf.t -> exception_value -> unit

Output a JSON value of type exception_value.

val string_of_exception_value : ?⁠len:int -> exception_value -> string

Serialize a value of type exception_value into a JSON string.

parameter len

specifies the initial length of the buffer used internally. Default: 1024.

val read_exception_value : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> exception_value

Input JSON data of type exception_value.

val exception_value_of_string : string -> exception_value

Deserialize JSON data of type exception_value.

val write_exception_ : Bi_outbuf.t -> exception_ -> unit

Output a JSON value of type exception_.

val string_of_exception_ : ?⁠len:int -> exception_ -> string

Serialize a value of type exception_ into a JSON string.

parameter len

specifies the initial length of the buffer used internally. Default: 1024.

val read_exception_ : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> exception_

Input JSON data of type exception_.

val exception__of_string : string -> exception_

Deserialize JSON data of type exception_.

val write_datetime : Bi_outbuf.t -> datetime -> unit

Output a JSON value of type datetime.

val string_of_datetime : ?⁠len:int -> datetime -> string

Serialize a value of type datetime into a JSON string.

parameter len

specifies the initial length of the buffer used internally. Default: 1024.

val read_datetime : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> datetime

Input JSON data of type datetime.

val datetime_of_string : string -> datetime

Deserialize JSON data of type datetime.

val write_breadcrumb : Bi_outbuf.t -> breadcrumb -> unit

Output a JSON value of type breadcrumb.

val string_of_breadcrumb : ?⁠len:int -> breadcrumb -> string

Serialize a value of type breadcrumb into a JSON string.

parameter len

specifies the initial length of the buffer used internally. Default: 1024.

val read_breadcrumb : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> breadcrumb

Input JSON data of type breadcrumb.

val breadcrumb_of_string : string -> breadcrumb

Deserialize JSON data of type breadcrumb.

val write_event : Bi_outbuf.t -> event -> unit

Output a JSON value of type event.

val string_of_event : ?⁠len:int -> event -> string

Serialize a value of type event into a JSON string.

parameter len

specifies the initial length of the buffer used internally. Default: 1024.

val read_event : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> event

Input JSON data of type event.

val event_of_string : string -> event

Deserialize JSON data of type event.